Refactor remaining routes to controllers
except for two relating to auth
